+++ frontend/src/addons/volto-frontend/src/constants/ActionTypes.js
@@ -1,1 +1,2 @@
 export const GENERATE = 'GENERATE';
+export const CHAT = 'CHAT';