+++ frontend/src/addons/volto-frontend/src/reducers/index.js
@@ -1,6 +1,8 @@
 import generate from './generate/generate';
+import chat from './chat/chat';
 
 const reducers = {
+  chat,
   generate,
 };